# [FrameworkBundle] Optimize framework extension tests