--- a/src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/FrameworkExtensionTest.php
+++ b/src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/FrameworkExtensionTest.php
@@ -312,7 +312,6 @@ abstract class FrameworkExtensionTest extends TestCase
         $dataHash = md5(serialize($data));
 
         if (!isset(self::$containerCache[$file][$dataHash])) {
-
             $container = $this->createContainer($data);
 
             $container->registerExtension(new FrameworkExtension());