@javiereguiluz 2016-01-16
@paradajozsef this is really impressive. This change keeps code readable and, according to your stats, it saves around 20 seconds in total. Thanks for this contribution and for your perfectly explained pull request.