@paradajozsef 2016-01-16
@javiereguiluz thank you. :) renamed ```$dataHash``` to ```$cacheKey```.