@paradajozsef 2016-01-17
@fabpot Yes, it should have been done so for the first time. Anyway, I changed the if statement.